backend/requirements.txt
fastapi
uvicorn
torch
torchvision
segmentation-models-pytorch
pillow
numpy
opencv-python
scikit-image
scipy
